+++ spec/wavefront-sdk/alert_spec.rb
@@ -14,6 +14,17 @@ ALERT_BODY = {
   severity:            'INFO'
 }.freeze
 
+def search_body(val)
+  { limit: 999,
+    offset: 0,
+    query: [
+      { key: 'status',
+        value: val,
+        matchingMethod: 'EXACT' }
+    ],
+    sort: { field: 'status', ascending: true } }
+end
+
 # Unit tests for Alert class
 #
 class WavefrontAlertTest < WavefrontTestBase
@@ -100,5 +111,67 @@ class WavefrontAlertTest < WavefrontTestBase
 
   def test_summary
     should_work(:summary, nil, 'summary')
+  end
+
+  def test_alerts_in_state
+    should_work(:alerts_in_state, 'some_state', '/api/v2/search/alert',
+                :post, {}, search_body('some_state'))
+  end
+
+  def test_firing
+    should_work(:firing, nil, '/api/v2/search/alert', :post, {},
+                search_body('firing'))
+  end
+
+  def test_active
+    should_work(:firing, nil, '/api/v2/search/alert', :post, {},
+                search_body('firing'))
+  end
+
+  def test_affected_by_maintenance
+    should_work(:affected_by_maintenance, nil, '/api/v2/search/alert',
+                :post, {}, search_body('in_maintenance'))
+  end
+
+  def test_invalid
+    should_work(:invalid, nil, '/api/v2/search/alert', :post, {},
+                search_body('invalid'))
+  end
+
+  def test_in_maintenance
+    should_work(:affected_by_maintenance, nil, '/api/v2/search/alert',
+                :post, {}, search_body('in_maintenance'))
+  end
+
+  def test_none
+    should_work(:none, nil, '/api/v2/search/alert', :post, {},
+                search_body('none'))
+  end
+
+  def test_checking
+    should_work(:checking, nil, '/api/v2/search/alert', :post, {},
+                search_body('checking'))
+  end
+
+  def test_trash
+    should_work(:trash, nil, '/api/v2/search/alert', :post, {},
+                search_body('trash'))
+  end
+
+  def test_no_data
+    should_work(:no_data, nil, '/api/v2/search/alert',
+                :post, {}, search_body('no_data'))
+  end
+
+  def test_snoozed
+    should_work(:snoozed, nil, '/api/v2/search/alert',
+                :post, {}, search_body('snoozed'))
+  end
+
+  # Not a full test, because it's a recursive API call. Just test
+  # the first API call is made correctly.
+  #
+  def test_all
+    should_work(:all, nil, '?offset=0&limit=999')
   end
 end